% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_grade.R
\name{compute_grade}
\alias{compute_grade}
\title{Calculate percentages, weighted averages and total grade}
\usage{
compute_grade(x, drop_n = NULL)
}
\arguments{
\item{x}{A \code{tbl_prof} object.}

\item{drop_n}{A numeric vector specifying whether lowest scores should be dropped in specific evaluation types.}
}
\value{
A \code{tbl_prof} object.
}
\description{
The function computes the percentages, weighted averages and the total grade.
Scores to drop (lowest) can also be specified.
}
