% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_groups.R
\name{make_groups}
\alias{make_groups}
\title{Make groups}
\usage{
make_groups(.data, group_size, alt_group_size = "larger",
  groups_col = NULL)
}
\arguments{
\item{.data}{A 2-dimensional structure (e.g. data frame, tibble, data.table,
...) that contains information about the population to be assigned into
groups.}

\item{group_size}{An integer specifying the desired group size.}

\item{alt_group_size}{A character value taking \code{"smaller"} or
\code{"larger"} to specify whether smaller or larger groups should be
formed in case of imperfect scenarios.}

\item{groups_col}{A character value specifying whether a column in
\code{.data} already exist. This is useful if the data already has
observations already assigned to groups groups. In this case, the
observations with no group should have \code{NA}.}
}
\description{
Make groups by specifying the population size, the group size and other
parameters.
}
