% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_data.R
\docType{data}
\name{scores}
\alias{scores}
\title{Evaluation scores for students.}
\format{A tibble (\code{tbl}) of 30 rows and 11 columns:
\describe{
    \item{student}{first name, last name}
    \item{quiz_1, quiz_2, quiz_3}{quiz scores out of 20}
    \item{homework_1, homework_2, homework_3}{homework scores out of 50}
    \item{exam_1, exam_2, exam_3, final}{exam scores out of 100}#' 
}}
\usage{
scores
}
\description{
A dataset containing scores on quizzes, homework assignments, exams and
attendance for a hypothetical class of 30 students.
}
\keyword{datasets}
