% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_prof.R
\name{tbl_prof}
\alias{tbl_prof}
\title{Create a tbl_prof object}
\usage{
tbl_prof(x, evals, max_scores = NULL, weights = NULL)
}
\arguments{
\item{x}{A data frame.}

\item{evals}{A character vector of regular expressions used to identify evaluation types in \code{x}.}

\item{max_scores}{A numeric vector specifying the maximum scores for each evaluation type.}

\item{weights}{A numeric vector specifying the weights of each evaluation type. Weights are decimal numbers between 0 and 1.}
}
\value{
A tbl_prof object.
}
\description{
Creating a \code{tbl_prof} object marks the start of the workflow.
Specify evaluation types, maximum scores and weights.
}
