% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_letter_grade.R
\name{assign_letter_grade}
\alias{assign_letter_grade}
\title{Assign letter grades}
\usage{
assign_letter_grade(x, intervals = NULL, letter_grades = NULL)
}
\arguments{
\item{x}{A \code{tbl_prof} object.}

\item{intervals}{A numeric vector specifying intervals for grades.}

\item{letter_grades}{A character vector of letter grades.}
}
\value{
a \code{tbl_prof} object.
}
\description{
Assign a letter grade based on the grade computed by the \code{compute_grade} function.
}
