% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_prof.R
\name{tbl_prof}
\alias{tbl_prof}
\title{Create a tbl_prof object}
\usage{
tbl_prof(x, evals, max_scores = NULL, weights = NULL)
}
\arguments{
\item{x}{A data frame containing evaluation scores.}

\item{evals}{A character vector of regular expressions used to identify evaluation types in \code{x}.}

\item{max_scores}{A numeric vector specifying the maximum scores for each evaluation type.}

\item{weights}{A numeric vector specifying the weights of each evaluation type. Each weight is a decimal number most likely between 0 and 1.}
}
\value{
A \code{tbl_prof} object.
}
\description{
The function creates a \code{tbl_prof} object. This is the first step of the workflow as all subsequent
operations can only be performed on a \code{tbl_prof} object.
}
